+++ fused.h
@@ -1,3 +1,5 @@
 #pragma once
 
-void fused(int* dst, const int* src, const char* cov, int n);
+#include <stddef.h>
+
+void fused(int* dst, const int* src, const char* cov, size_t n);
